+++ tests/Equinox.EventStoreDb.Integration/Infrastructure.fs
@@ -3,6 +3,7 @@
 open Domain
 open FsCheck.FSharp
 open System
+open System.Diagnostics
 open FSharp.UMX
 
 module ArbMap =
@@ -30,6 +31,22 @@ open Equinox.EventStoreDb
 open Equinox.EventStore
 #endif
 
+[<RequireQualifiedAccess>]
+type EsAct = Append | AppendConflict | SliceForward | SliceBackward | BatchForward | BatchBackward | ReadLast
+
+[<AutoOpen>]
+module OtelHelpers =
+    let (|EsOp|_|) (act: Activity): EsAct option =
+        match act.OperationName with
+        | "Append" -> Some EsAct.Append
+        | "AppendConflict" -> Some EsAct.AppendConflict
+        | "SliceForward" -> Some EsAct.SliceForward
+        | "SliceBackward" -> Some EsAct.SliceBackward
+        | "BatchForward" -> Some EsAct.BatchForward
+        | "BatchBackward" -> Some EsAct.BatchBackward
+        | "ReadLast" -> Some EsAct.ReadLast
+        | _ -> None
+
 [<AutoOpen>]
 module SerilogHelpers =
     open Serilog.Events
@@ -37,8 +54,6 @@ module SerilogHelpers =
     let (|SerilogScalar|_|): LogEventPropertyValue -> obj option = function
         | :? ScalarValue as x -> Some x.Value
         | _ -> None
-    [<RequireQualifiedAccess>]
-    type EsAct = Append | AppendConflict | SliceForward | SliceBackward | BatchForward | BatchBackward | ReadLast
     let (|EsAction|) (evt: Log.Metric) =
         match evt with
         | Log.WriteSuccess _ -> EsAct.Append
@@ -67,6 +82,52 @@ module SerilogHelpers =
     let (|SerilogString|_|): LogEventPropertyValue -> string option = function SerilogScalar (:? string as y) -> Some y | _ -> None
     let (|SerilogBool|_|): LogEventPropertyValue -> bool option = function SerilogScalar (:? bool as y) -> Some y | _ -> None
 
+/// Captures store-specific Activity emissions for test assertions, using AsyncLocal for per-test isolation
+type StoreActivityCapture() =
+    static let storeSourceName =
+#if STORE_EVENTSTOREDB
+        "Equinox.EventStoreDb"
+#endif
+#if STORE_EVENTSTORE_LEGACY
+        "Equinox.EventStore"
+#endif
+#if STORE_MESSAGEDB
+        "Equinox.MessageDb"
+#endif
+#if STORE_POSTGRES || STORE_MSSQL || STORE_MYSQL
+        "Equinox.SqlStreamStore"
+#endif
+    static let currentOps = new System.Threading.AsyncLocal<ResizeArray<string>>()
+    static let gate = obj()
+    static let _listener =
+        let l = new ActivityListener(
+            ShouldListenTo = (fun source -> source.Name = storeSourceName),
+            Sample = (fun _ -> ActivitySamplingResult.AllDataAndRecorded),
+            ActivityStopped = (fun act ->
+                let ops = currentOps.Value
+                if ops <> null then lock gate (fun () -> ops.Add(act.OperationName))))
+        ActivitySource.AddActivityListener(l)
+        l
+    let ops = ResizeArray<string>()
+    do
+        ignore _listener
+        currentOps.Value <- ops
+    member _.Clear() = lock gate (fun () -> ops.Clear())
+    member _.ExternalCalls =
+        lock gate (fun () -> ops |> Seq.toList)
+        |> List.choose (fun name ->
+            match name with
+            | "Append" -> Some EsAct.Append
+            | "AppendConflict" -> Some EsAct.AppendConflict
+            | "SliceForward" -> Some EsAct.SliceForward
+            | "SliceBackward" -> Some EsAct.SliceBackward
+            | "BatchForward" -> Some EsAct.BatchForward
+            | "BatchBackward" -> Some EsAct.BatchBackward
+            | "ReadLast" -> Some EsAct.ReadLast
+            | _ -> None)
+    interface IDisposable with
+        member _.Dispose() = currentOps.Value <- null
+
 type LogCapture() =
     inherit LogCaptureBuffer()
     member _.ExternalCalls = base.ChooseCalls(function EsEvent (EsAction act) -> Some act | _ -> None)
@@ -74,6 +135,17 @@ type LogCapture() =
 type TestContext(testOutputHelper) =
     let output = TestOutput testOutputHelper
 
-    member x.CreateLoggerWithCapture() =
+    member _.CreateLogger() = output.CreateLogger()
+
+    member _.CreateLoggerWithCapture() =
+        let capture = new StoreActivityCapture()
+        output.CreateLogger(), capture
+
+    member _.CreateLoggerWithSerilogCapture() =
         let capture = LogCapture()
         output.CreateLogger(capture), capture
+
+    member _.CreateLoggerWithDualCapture() =
+        let serilogCapture = LogCapture()
+        let otelCapture = new StoreActivityCapture()
+        output.CreateLogger(serilogCapture), otelCapture, serilogCapture